Add Cell tests and fix hash randomisation
Games::TMX::Parser::Cell objects relied on a consistent
hash key order, which is no longer guarranteed after Perl 5.12.

This reimplements the seek_next_cell method so it will continue
to work on newer versions of Perl, and adds tests for that and
other Cell methods, which were not being covered.